+++ Source/Bumpy/Config/ConfigIO.cs
@@ -1,23 +1,31 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System.Collections.Generic;
+using System.IO;
 using System.Text;
-using System.Threading.Tasks;
+using YamlDotNet.Serialization;
 
 namespace Bumpy.Config
 {
+    // TODO tests
     public static class ConfigIO
     {
         public static string NewConfigFile()
         {
-            // TODO fw
-            return "# Bumpy configuration file";
+            var builder = new StringBuilder();
+            builder.AppendLine("# Bumpy configuration file");
+            builder.AppendLine("- glob: \"**\\AssemblyInfo.cs\"");
+            builder.AppendLine("  profile: assembly");
+            builder.AppendLine();
+            builder.AppendLine("- glob: \"**\\*.nuspec\"");
+            builder.AppendLine("  profile: nuget");
+
+            return builder.ToString();
         }
 
-        // TODO fw Template Bestimmung und Profil Einschränkung könnte ich für beide ConfigIOs zusammenfassen
-        public static IEnumerable<BumpyConfigEntry> ReadConfigFile(IEnumerable<string> lines, string profile)
+        public static IEnumerable<BumpyConfigEntry> ReadConfigFile(TextReader reader)
         {
-            return null;
+            var deserializer = new Deserializer();
+
+            return deserializer.Deserialize<List<BumpyConfigEntry>>(reader);
         }
     }
 }